Complete RC-20

+++ core/src/main/java/com/dragonentertainment/runningcat/utils/GameGrid.java
@@ -6,7 +6,7 @@ public class GameGrid {
     public static final int CELL_SIZE = 1;
     public static final int WORLD_WIDTH = 9;
     public static final int WORLD_HEIGHT = 16;
-    private static final float PIXEL_PER_UNIT = Gdx.graphics.getHeight() / WORLD_HEIGHT;
+    private static final float PIXEL_PER_UNIT = (float) Gdx.graphics.getHeight() / WORLD_HEIGHT;
 
     public static int toGridWidth(float textureWith) {
         return (int) (textureWith / PIXEL_PER_UNIT);
